Prevent long-range birthday attacks on knockout merkles by injecting blockhash entropy
The pre-existing Merkle proof for knockout history used the timestamp and pool's fee
accumulator. Theoritically both of these components could be manipulated to allow for
a long-range attack. Since the Merkle hash is 160-bits, a possible birhtday attack
collission could be carried out with $10 billion of hash power.

This update eliminates the ability of an attacker to utilize a long-range birthday
collission. It injects the leading 160 bits of the block hash into the Merkle hash, which
prevents a long-range attacker from being able to fully predict/manipulate the proof
ahead of time.

+++ contracts/callpaths/KnockoutPath.sol
@@ -145,7 +145,7 @@ contract KnockoutLiqPath is TradeMatcher, SettleLayer {
                        KnockoutLiq.KnockoutPosLoc memory loc,
                        bytes memory args) private returns
         (int128 baseFlow, int128 quoteFlow) {
-        (uint160 root, uint96[] memory proof) = abi.decode(args, (uint160,uint96[]));
+        (uint160 root, uint256[] memory proof) = abi.decode(args, (uint160,uint256[]));
         (baseFlow, quoteFlow) = claimKnockout(curve, loc, root, proof, pool);
         commitCurve(pool, curve);
     }
